third_party/blink/web_tests/external/wpt/css/mediaqueries/inverted-colors-expected.txt

This is a testharness.js-based test.
FAIL Should be known: '(inverted-colors)' assert_true: expected true got false
FAIL Should be known: '(inverted-colors: none)' assert_true: expected true got false
FAIL Should be known: '(inverted-colors: inverted)' assert_true: expected true got false
PASS Should be parseable: '(inverted-colors: 0)'
PASS Should be unknown: '(inverted-colors: 0)'
PASS Should be parseable: '(inverted-colors: no-preference)'
PASS Should be unknown: '(inverted-colors: no-preference)'
PASS Should be parseable: '(inverted-colors: 10px)'
PASS Should be unknown: '(inverted-colors: 10px)'
PASS Should be parseable: '(inverted-colors: none inverted)'
PASS Should be unknown: '(inverted-colors: none inverted)'
PASS Should be parseable: '(inverted-colors: none/inverted)'
PASS Should be unknown: '(inverted-colors: none/inverted)'
FAIL Check that none evaluates to false in the boolean context assert_equals: expected true but got false
PASS Check that invalid evaluates to false
Harness: the test ran to completion.
